src/user/domain/user/user.ts
export class User {
  private email: string;
  private password: string;

  constructor(
    public readonly id: string,
    private name: string,
    private readonly createdAt: string,
    private readonly updatedAt: string,
  ) {}

  public getEmail(): string {
    return this.email;
  }

  public setEmail(email: string): void {
    this.email = email;
  }

  public getPassword(): string {
    return this.password;
  }

  public setPassword(password: string): void {
    this.password = password;
  }

  public getName(): string {
    return this.name;
  }

  public setName(name: string): void {
    this.name = name;
  }

  public getCreatedAt(): string {
    return this.createdAt;
  }

  public getUpdatedAt(): string {
    return this.updatedAt;
  }
}
